Replace state setter in context with custom setter function
This function takes the property name as the first argument, and the value to set it to as the second argument. This means I no longer have to spread the existing session state into a new object before setting the selected property's new value; instead I can just do that in the actual state setter within the LootGenerator component.