Write initial loot generation functions
Currently the loot generation is very primitive; it simply loops through a provided table the specified number of times and searches through the table and any sub-tables, selecting items at random based on their weightings. It doesn't yet factor in the 'rolls' field of the GenerationCriteria type - this is something I will attempt to add in future.