Remove 'LootTableBase' type and associated functions
It wasn't really necessary to have this type; the base loot table should have access to all the same properties as a normal loot table. The only slightly annoying thing is that the base loot table must have a 'weight' (as any other loot table must have one) even though it won't be checked as the base loot table isn't rolled, but this is a minor inconvenience.