Adjust layout to remove 'empty' space around menu content
Due to content not occupying the full space of the bar as a result of padding and automatic sizing, there were spaces where, when hovering over the menu but not the elements within it, the hover styles weren't being applied as they are triggered when hovering the elements within the bar and not the bar itself. They now occupy all the available space of the bar which has fixed the issue.