Create Preset type
This type is identical to either a LootItem or LootTable, with the 'key' field missing. This is because presets will be stored in a Map for efficient lookup; when a preset is included in the active loot table, I will check to see if any entry's key matches a preset and load the information for that preset if applicable.