Remove Preset type and replace presets state with array
Even though it's not as efficient to loop through an array looking for a preset, this approach will allow me to reuse existing components for the interactive view for both the Active and Presets tabs. I am also considering creating a Map containing every presets for quick lookup that just gets updated every time a preset is created/mutated/deleted.