Move the Presets tab to be a sub-tab of the Interactive tab
I have also provided it with the same 'menuStates' context as the Active sub-tab, which should allow me to control which entries are open/closed for each sub-tab independently. The context now also contains a 'menuType' that is different for the Active and Presets sub-tabs, which will allow me to control which elements are visible in TableEntry and ItemEntry components, for example, the SaveAsPresetButton component shouldn't be present on an entry in the Presets sub-tab.