Directly replace entry in LootTable 'props.loot' array
I was trying to replace an entry's fields with that of a preset before, which felt like bad practise - it seemed better to just replace the entry entirely with a LootPreset. I have also deleted the 'convertEntryToLootPreset' function and have prevented the active lootTable state from being saved as a preset.