Make sure default styles object is copied before merge
Lodash's 'merge' function was causing duplicate styles across multiple instances of the ToggleBar component; creating a copy of the default styles object before merging it with the 'style' prop prevents this. I think this problem arose because all instances of the components shared the same default styles object, and therefore shared the same styles as the last instance of the component to be mounted, as the object reference would have been identical and the last instance to be mounted would have been the last thing to mutate this object.