Remove ability to make item presets
As items are now being created in a separate tab, I am planning to give item entries in loot tables an 'id' field referencing an item, which reduces the need to have item presets. The generation criteria for items will still be defined within the item entry in a loot table and not on the item itself, so there may be some benefit to item presets, for example if the user wishes to have many variations of an item, but I will implement this in future if I decide this is necessary.