Remove 'menu states' from Interactive component
I'm going to aim to find a better way to achieve state maintenance for the tab components which are unmounted when their tab is deselected - the current method I was using by saving the toggle state for each ToggleBar component was causing issues with multiple components using the same key within the 'menuStates' map, resulting in them both being open/closed together. Also, when the Interactive tab itself was closed, all the ToggleBar default states were reset anyway, so it wasn't a great solution. I imagine some form of memoisation or lifting the components to a higher-level component would be a more effective solution.