Change the way tables are handled to be more similar to items
This means displaying the top-level tables in the Tables tab using the Table component, and all descendant table/item entries in the 'loot' array using the LootTable and LootItem components, respectively. The TableEntry now displays a SelectTable component, from which a table can be selected (although this rework has broken a few of the LootGenerator context functions, which I will rectify soon). All descendant entries are displayed within the Table component, rather than within the TableEntry component - the latter now only displays fields not bound to a defined table, such as the generation criteria fields, similar to item entries. This also solves the problem of having unnecessary fields on the top-level tables; for example, they do not need the generation criteria fields.