Update LootTable and LootItem types
To allow tables and items with and without specified ids, I have updated the 'type' field to allow both 'table_id' and 'table_noid' for the LootTable type, and 'item_id' and 'item_noid' for the LootItem type. This type will dictate which fields should be present on the LootTable/LootItem - where an 'id' field should not be specified, the table/item's fields will be defined directly on the LootTable and LootItem objects. I obviously need to update some functions and potentially write new ones to allow switching between the two states for both tables and items, but I think this is a good solution for the issue of wanting to have editable table and item entries without having to define their properties outside of a table first (although I reserve the right to change 'xxxx_id'/'xxxx_noid' as identifiers later).