Write 'displayPopUpMessage' function
I had to use a workaround for getting the state value containing the PopUpModal components to actually have the correct messages displayed. I couldn't mutate that exact state value from within the callback passed to the PopUpModal components, but I could store the components in a ref and mutate a different state value. The mutation of this state is observed by a useMemo hook which creates the JSX element containing the PopUpModal components.