Remove array filter when displaying tables in SelectEntry component
This doesn't currently work as intended; it doesn't accurately check if a table to be selected as an entry within another table's loot array is already an ancestor table of that table, which would cause an infinite recursion of tables referencing each other. I need to implement a more robust solution to this issue.