Fix setIdOnEntry function
The function wasn't adding new tables/items to the LootGenerator component's state when 'noid' tables/items had ids set on them. Also, the type checking at the start of the function was wrong; I needed to check the 'type' field of the specified entry, not whether the new id existed in the LootGenerator component's state's tables/items.