Write findCompatibleDescendantTables function
This function scans all tables in the LootGenerator component's state, and finds which of them can be set as a descendant table of the specified table. If a table is an ancestor of the specified table at any point, it cannot be set as a descendant table, otherwise the possibility of an infinite recursion arises when generating loot from the table.